+++ src/aux-common/runtime/AuxDevice.ts
@@ -11,4 +11,12 @@ export interface AuxDevice {
      * Whether the device supports virtual reality features.
      */
     supportsVR: boolean;
+
+    /**
+     * Whether this device has enabled collaboration features.
+     *
+     * When creating a simulation, this property may be used to enable or disable features that facilitate users interacting with each other.
+     * For example, setting isCollaborative to false would make the shared partition act like a tempLocal partition.
+     */
+    isCollaborative: boolean;
 }