+++ src/aux-server/shared/WebConfig.ts
@@ -28,4 +28,11 @@ export interface WebConfig {
      * The HTTP Origin that should be used for VM Iframes.
      */
     vmOrigin?: string;
+
+    /**
+     * Whether collaboration should be disabled.
+     * Setting this to true will replace the shared partition of simulations
+     * with tempLocal partitions.
+     */
+    disableCollaboration?: boolean;
 }