+++ src/aux-vm/vm/BaseAuxChannel.ts
@@ -275,6 +275,7 @@ export abstract class BaseAuxChannel implements AuxChannel, SubscriptionLike {
         this._statusHelper = new StatusHelper(
             partitions.map((p) => p.onStatusUpdated)
         );
+        this._statusHelper.defaultUser = this.user;
 
         let statusMapper = remapProgressPercent(0.3, 0.6);
         this._subs.push(
@@ -491,7 +492,6 @@ export abstract class BaseAuxChannel implements AuxChannel, SubscriptionLike {
             message: 'Launching interface...',
             progress: 0.9,
         });
-        await this._initUserDimensionBot();
 
         if (!this._hasRegisteredSubs) {
             this._hasRegisteredSubs = true;
@@ -645,17 +645,6 @@ export abstract class BaseAuxChannel implements AuxChannel, SubscriptionLike {
             await this._helper.createOrUpdateUserBot(this.user, userBot);
         } catch (err) {
             console.error('[BaseAuxChannel] Unable to init user bot:', err);
-        }
-    }
-
-    private async _initUserDimensionBot() {
-        try {
-            await this._helper.createOrUpdateUserDimensionBot();
-        } catch (err) {
-            console.error(
-                '[BaseAuxChannel] Unable to init user dimension bot:',
-                err
-            );
         }
     }